refactor: restructure services/common to split alloy & axum layers
